heat implementation updated
